remote_monitoring-server/requirements.txt
aioredis==1.3.1
appnope==0.1.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
backcall==0.1.0
cffi==1.14.0
channels==2.0.0
channels-redis==2.2.1
constantly==15.1.0
cryptography==2.8
decorator==4.4.2
Django==2.0
django-cors-headers==3.2.1
django-taggit==1.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
msgpack==0.5.6
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2-binary==2.7.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pytz==2019.3
service-identity==18.1.0
six==1.14.0
traitlets==4.3.3
Twisted==20.3.0
txaio==20.3.1
wcwidth==0.1.9
zope.interface==5.0.2
